Refactor WordHighlighter to use custom hook for improved readability and maintainability
