+++ src/features/lireFlow/components/WordHighlighter.tsx
@@ -1,7 +1,5 @@
-import { useEffect, useState, useRef, RefObject } from "react";
-import { hyphenate } from "hyphen/pt";
-import { averageSyllableTime } from "./ReadingParameters";
-import { punctuationMarksTime } from "./ReadingParameters";
+import React, { RefObject } from "react";
+import { useWordHighlighter } from "../hooks/useWordHighlighter"; 
 
 const WordHighlighter = ({
   paragraph,
@@ -16,93 +14,13 @@ const WordHighlighter = ({
   speedRef: RefObject<number>;
   wordsPerMinuteRef: RefObject<number>;
 }) => {
-  const [currentIndex, setCurrentIndex] = useState<number | null>(null);
-  const elementIndexs = useRef<number[]>([]);
-  const indexRef = useRef<number>(0);
-  const isReadingRef = useRef<boolean>(isReading);
-  const timeoutRef = useRef<number | null>(null);
-  const currentWordRef = useRef<HTMLSpanElement | null>(null);
-
-  const elements: string[] = paragraph.split(/(\s+|[^\wÀ-ÖØ-öø-ÿ])/);
-
-  const isWord = (element: string): boolean => /^[\wÀ-ÖØ-öø-ÿ]+$/.test(element);
-  const isPunctuation = (element: string): boolean =>
-    /^[.,!?;:"()]+$/.test(element);
-
-  useEffect((): void => {
-    elementIndexs.current = elements
-      .map((element: string, index: number): number | null =>
-        isWord(element) || isPunctuation(element) ? index : null
-      )
-      .filter((index): index is number => index !== null);
-
-    indexRef.current = 0;
-    setCurrentIndex(null);
-  }, [paragraph]);
-
-  useEffect((): (() => void) => {
-    isReadingRef.current = isReading;
-
-    const calculateWordTime = async (word: string): Promise<number> => {
-      const hyphenatedText: string = await hyphenate(word, { hyphenChar: "-" });
-      const syllablesCount: number = hyphenatedText.split("-").length;
-      return Math.round(
-        (syllablesCount * averageSyllableTime(wordsPerMinuteRef.current)) /
-          speedRef.current
-      );
-    };
-
-    const highlightFlow = async (): Promise<void> => {
-      while (indexRef.current < elementIndexs.current.length) {
-        if (!isReadingRef.current) return;
-
-        const element: string =
-          elements[elementIndexs.current[indexRef.current]];
-
-        let waitTime: number = 0;
-
-        if (isWord(element)) {
-          setCurrentIndex(elementIndexs.current[indexRef.current]);
-          waitTime = await calculateWordTime(element);
-        } else {
-          waitTime =
-            punctuationMarksTime.find((mark) => mark.mark === element)?.time ||
-            150;
-        }
-
-        await new Promise(
-          (resolve) => (timeoutRef.current = setTimeout(resolve, waitTime))
-        );
-
-        indexRef.current++;
-      }
-
-      setCurrentIndex(null);
-      if (onFinish) {
-        await new Promise(
-          (resolve) => (timeoutRef.current = setTimeout(resolve, 500))
-        );
-        onFinish();
-      }
-    };
-
-    highlightFlow();
-
-    return () => {
-      if (timeoutRef.current) {
-        clearTimeout(timeoutRef.current);
-      }
-    };
-  }, [isReading]);
-
-  useEffect(() => {
-    if (currentWordRef.current) {
-      currentWordRef.current.scrollIntoView({
-        behavior: "smooth",
-        block: "nearest",
-      });
-    }
-  }, [currentIndex]);
+  const { elements, currentIndex, currentWordRef } = useWordHighlighter({
+    paragraph,
+    onFinish,
+    isReading,
+    speedRef,
+    wordsPerMinuteRef,
+  });
 
   return elements.map((element: string, index: number) => (
     <span
